pyproject.toml
[project]
name = "stt-vosk"
version = "0.1.0"
description = "A modular service that provides speech-to-text (STT) capabilities for machines running on the Viam platform."
authors = [
    { name = "HipsterBrown", email = "headhipster@hipsterbrown.com" }
]
dependencies = [
    "viam-sdk>=0.15.0",
    "vosk>=0.3.44",
    "speechrecognition>=3.10.1",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
]
readme = "README.md"
requires-python = ">= 3.8"
license = { text = "Apache-2.0" }

[project.scripts]
"stt-vosk" = "stt_vosk:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/stt_vosk"]
